added safety assert to inventory constructor, made Remove function more readable, and made naming more consistent on variables

+++ Code/Inventory.h
@@ -6,7 +6,7 @@
 class Inventory
 {
 public:
-    Inventory(short InCapacity);
+    Inventory(size_t InCapacity);
 
     int AddItem(const Item& ItemToAdd, int Quantity);
